Code cleanup, don't use deprecated mgmt apis

+++ testsuite/domain/src/test/java/org/jboss/as/test/integration/domain/management/cli/DeploySingleServerGroupTestCase.java
@@ -76,7 +76,9 @@ public class DeploySingleServerGroupTestCase extends AbstractCliTestBase {
     @AfterClass
     public static void after() throws Exception {
         AbstractCliTestBase.closeCLI();
-        Assert.assertTrue(warFile.delete());
+        if (warFile!=null) {
+            Assert.assertTrue(warFile.delete());
+        }
 
         CLITestSuite.stopSupport();
     }
